requirements.txt
Flask==3.1.0
flask_cors==5.0.1
fpdf2==2.8.2
numpy==1.24.4
Pillow==11.1.0
python-dotenv==1.1.0
torch==2.2.2+cpu
ultralytics==8.3.98
unstract==0.0.1
gunicorn
llmwhisperer-client==2.2.1  # o la última versión estable
opencv_contrib_python==4.8.0.76
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
--extra-index-url https://download.pytorch.org/whl/cpu